Actually move the unit tests to the same project as the source
