+++ atlasdb-cassandra/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraKeyValueServiceConfigTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.atlasdb.keyvalue.cassandra.unittest;
+package com.palantir.atlasdb.keyvalue.cassandra;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
